### 2024-11-12 #7469 @editorialbot
[PRE REVIEW]: ORC: The Open Research Converter
**Submitting author:** @jhculb (<a href="http://orcid.org/0009-0000-1581-4021">Jack H. Culbert</a>)
**Repository:** https://github.com/jhculb/Open-Research-Converter
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
