### 2024-11-12 #7469 @samhforbes
Hi @jhculb thanks for your submission to JOSS.
Given the relatively small size of this package it will be first checked for scope by our editors.
