### 2025-01-06 #7469 @logological
@samhforbes I'm interested in the paper but I'm currently at capacity.  I could take it on as soon as one of my existing assignments is completed.
