### 2025-01-07 #7469 @logological
@jhculb I'd be delighted to begin handling this submission as soon as one of my existing assignments is completed.  In the meantime, please let me know if you have any recommendations for reviewers.  If they have GitHub accounts, you could mention those here (without using an @-mention).  You might also search the [the list of people](https://reviewers.joss.theoj.org/lookup) who have already agreed to review and may be suitable for this submission.
