### 2025-02-05 #7469 @jhculb
Hi @logological, thanks for agreeing to handle this!

I think Nolan McCafferty - NolanJMcCafferty - developer at OurResearch may be particularly well placed for review of the OpenAlex API and Scientometrics.

Otherwise from the JOSS reviewers list:

- cerule7
- sarahsid10
- gagolews
