### 2025-04-25 #7469 @logological
@NolanJMcCafferty, @gagolews, @cerule7, and @sarahsid10, would any of you be willing and able to review this submission for the _Journal of Open Source Software_?  It introduces ORC, a tool for bulk conversion of bibliometric data to a shareable format.
