### 2025-04-26 #7469 @cerule7
> [@NolanJMcCafferty](https://github.com/NolanJMcCafferty), [@gagolews](https://github.com/gagolews), [@cerule7](https://github.com/cerule7), and [@sarahsid10](https://github.com/sarahsid10), would any of you be willing and able to review this submission for the _Journal of Open Source Software_? It introduces ORC, a tool for bulk conversion of bibliometric data to a shareable format.

Hello @logological, it appears that you have tagged me in error. If this is an intentional ping, I am happy to give it a go :-)
