### 2025-04-28 #7469 @gagolews
> [@NolanJMcCafferty](https://github.com/NolanJMcCafferty), [@gagolews](https://github.com/gagolews), [@cerule7](https://github.com/cerule7), and [@sarahsid10](https://github.com/sarahsid10), would any of you be willing and able to review this submission for the _Journal of Open Source Software_? It introduces ORC, a tool for bulk conversion of bibliometric data to a shareable format.

Apologies, I'm currently too busy to handle this request.
