### 2025-04-28 #7469 @sarahsid10
Hello, I haven't had the chance to review JOSS submissions before but would be interested in trying. How soon would you like it to be done, and are there any guidelines/rubrics for evaluation?
