### 2025-04-28 #7469 @logological
@sarahsid Great that you're willing to give it a try!  We have a [reviewers' guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) that explains the procedure and should answer most or all of your questions.  (If you have any further questions, feel free to let me know.)  We generally ask reviewers to complete the reviews within four to six weeks.  Could you have a look at the guidelines and let me know if you think you'll be able to proceed?
