### 2025-04-28 #7469 @logological
@cerule7 Your comment seems to have disappeared; I'm not sure if that's because of some glitch or because you yourself deleted it.  In case it's the former, you were contacted because you appear in JOSS's Reviewer Management System as someone who had volunteered to consider reviewing for the journal for the topics "NLP, databases, cybersecurity, linguistics, software engineering".  If you don't remember signing up or are no longer interested in reviewing, let me know and I can see that you're removed from our database.  If you are still interested in serving as a reviewer, please let me know whether you'd be willing and able to review this particular submission.
