### 2025-05-15 #8205 @editorialbot
[REVIEW]: ORC: The Open Research Converter
**Submitting author:** @jhculb (<a href="http://orcid.org/0009-0000-1581-4021">Jack H. Culbert</a>)
**Repository:** https://github.com/jhculb/Open-Research-Converter
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.1
