### 2025-06-09 #8205 @dmnapolitano
Hi @jhculb! 👋🏻  I'm currently working on my review. Before I get to the code and trying to run it locally, here's some feedback:

Regarding documentation, would you consider adding API documentation should someone want to use ORC programmatically?

Some grammatical suggestions in the paper itself:
* On line 20, "to use the" should be "use of the".
* On line 33, "While it is possible" should be "The only options are".
* On lines 38-39, insert hyphens here: "Python-" and "Javascript-based", and also add commas, "which, when run, ..."

Regarding the web interface, I agree with @evamaxfield's feedback about informing the user of the number of articles that were found. If there's any information being returned by OpenAlex regarding the missing ones, that would be helpful to display, too (perhaps with a "verbose mode" the user can enable if this information is extremely, well, verbose 😄).

That's it for now! Thanks.
