### 2025-06-18 #8205 @logological
@sarahsid10 Just wanted to check in with you on the status of your review.  I note that you've still got some checklist items unchecked—have you just not got around to these yet, or have you discovered any problems with the paper or software that are preventing you from moving forward?  If the latter, it would be great if you could raise these problems here, or open individual issues for them in the Open-Research-Converter GitHub project.
