### 2025-07-16 b4c83f0
Add app docstrings
