### 2025-07-16 c8d77b6
Backend docstrings with process ordering
