### 2025-07-16 4543da7
Backend docstrings _check_ready
