### 2025-07-16 24ed7be
Class docstrings
