### 2025-08-02 #8205 @logological
@jhculb Have you been able to address the issues raised by the reviewers?  I see there's been some activity in the Open-Research-Converter repository in the last few weeks, but if I'm not mistaken this seems to be limited to updates to the documentation and paper.

@dmnapolitano, @evamaxfield, and @sarahsid10: Are the Docker installation issues some of you raised still blocking further progress on your reviews?
