### 2025-08-03 #8205 @jhculb
> @jhculb Have you been able to address the issues raised by the reviewers?  I see there's been some activity in the Open-Research-Converter repository in the last few weeks, but if I'm not mistaken this seems to be limited to updates to the documentation and paper.
>
> @dmnapolitano, @evamaxfield, and @sarahsid10: Are the Docker installation issues some of you raised still blocking further progress on your reviews?

Hello, unfortunately I have been hospitalised for the last month and unable to make major changes. I should be able to address the other points and the docker issue within the next 2 weeks.
