### 2025-12-07 1495b35
Add diun healthchecking and ignore database
Added diun configuration for container health monitoring and updated
gitignore to exclude diun database files.
