### 2026-01-27 134cf93
Add documentation and DOI tracking
