+++ src/orc/backend/orc_backend/open_research_converter.py
@@ -1,3 +1,28 @@
+"""
+Open Research Converter Core Module.
+
+This module contains the OpenResearchConverter class, which orchestrates the
+conversion of DOIs to OpenAlex identifiers. It handles job management, input
+validation, and coordinates with the OpenAlex API requester.
+
+The OpenResearchConverter is designed to support bibliometric research by enabling
+bulk conversion of DOIs from proprietary databases to the open OpenAlex format,
+facilitating reproducibility in scientometric studies.
+
+Classes:
+    OpenResearchConverter: Main class for managing DOI to OpenAlex ID conversion jobs.
+
+Example:
+    Basic usage::
+
+        from orc.backend.orc_backend.open_research_converter import OpenResearchConverter
+
+        orc = OpenResearchConverter(logger)
+        job_id = orc.generate_new_job()
+        await orc.process(job_id, ["10.1038/nature12373"], "researcher@university.edu")
+        result, status_code = orc.return_data(job_id)
+"""
+
 from __future__ import annotations
 
 import asyncio
@@ -8,15 +33,68 @@ from orc.backend.orc_backend.requester import OpenAlexRequester
 
 
 class OpenResearchConverter(OpenAlexRequester):
-    """This class contains code to interface between the frontend and the requester class, and orchestrate the processing of a job"""
+    """
+    Orchestrator for converting DOIs to OpenAlex identifiers.
+
+    This class manages the lifecycle of conversion jobs, including:
+    - Job creation and tracking
+    - Input validation (DOIs, email)
+    - Coordination with the OpenAlex API
+    - Result formatting and retrieval
+
+    Inherits from OpenAlexRequester to leverage async API communication capabilities.
+
+    Attributes:
+        _logger: Logger instance for debugging and error reporting.
+        _jobs (dict): Dictionary storing job data, keyed by job_id (UUID string).
+
+    Example:
+        Processing a list of DOIs::
+
+            orc = OpenResearchConverter(app.logger)
+            job_id = orc.generate_new_job()
+            await orc.process(
+                job_id,
+                "10.1038/nature12373, 10.1126/science.1231143",
+                "researcher@university.edu"
+            )
+            result, status = orc.return_data(job_id)
+            print(f"Found {result['found_count']} of {result['submitted_count']} DOIs")
+    """
 
     def __init__(self, log) -> None:
-        """Instantiate the Open Research Converter (functionality to serve the frontend) from the Requester class (functionality to communicate with OpenAlex)."""
+        """
+        Initialize the OpenResearchConverter.
+
+        Args:
+            log: Logger instance for logging debug information and errors.
+                Typically passed from the Flask/Quart application logger.
+        """
         super().__init__()
         self._logger = log
 
     def generate_new_job(self) -> str:
-        """1. Initialises the data structures associated with a job"""
+        """
+        Create a new processing job with a unique identifier.
+
+        Initializes all required data structures for tracking the job through
+        its lifecycle from creation to completion.
+
+        Returns:
+            str: A UUID string uniquely identifying this job.
+
+        Note:
+            The job is created with status "initialised" and must go through
+            the process() or process_all() method to be completed.
+
+        Job Data Structure:
+            - input_data: List of DOIs to process (None until data received)
+            - email: User's email for OpenAlex polite pool
+            - status: One of "initialised", "ready", "complete"
+            - submitted_count: Total DOIs submitted
+            - found_count: DOIs successfully matched in OpenAlex
+            - missing_dois: List of DOIs not found in OpenAlex
+        """
         new_job_id = uuid.uuid4().__str__()
         self._logger.debug(f"orc.py: new job created with id: {new_job_id}")
[…]
