+++ src/orc/backend/orc_backend/app.py
@@ -1,3 +1,37 @@
+"""
+Open Research Converter (ORC) API Module.
+
+This module provides the REST API endpoints for the Open Research Converter,
+a tool for converting DOIs (Digital Object Identifiers) to OpenAlex IDs.
+
+The API enables researchers to:
+- Convert lists of DOIs to OpenAlex identifiers
+- Retrieve full bibliometric metadata from OpenAlex
+- Check API health status
+
+Endpoints:
+    GET  /              : API information page
+    GET  /healthcheck   : Check OpenAlex API connectivity
+    POST /start_processing : Convert DOIs to OpenAlex IDs
+    POST /process_all   : Convert DOIs and retrieve full OpenAlex metadata
+
+Example:
+    To convert DOIs to OpenAlex IDs::
+
+        POST /start_processing
+        Content-Type: application/json
+        {
+            "email": "researcher@university.edu",
+            "input_data": "10.1038/nature12373, 10.1126/science.1231143"
+        }
+
+For more information, see the OpenAPI specification at openapi.yaml.
+
+References:
+    - OpenAlex API: https://docs.openalex.org/how-to-use-the-api/api-overview
+    - ORC Documentation: https://github.com/jhculb/OpenResearchConverter
+"""
+
 from __future__ import annotations
 
 import logging
@@ -5,8 +39,6 @@ import logging
 from orc.backend.orc_backend.open_research_converter import OpenResearchConverter
 from quart import Quart, jsonify, request
 from quart_cors import cors
-
-# from https://stackoverflow.com/questions/67741946/how-to-validate-fields-raw-in-flask-marshmallow?rq=1
 
 app = Quart(__name__)
 cors = cors(app)
@@ -22,7 +54,18 @@ orc = OpenResearchConverter(log)
 
 @app.route("/", methods=["GET"])
 def hello_world():
-    """Route stub for the API - useful for manual healthchecking"""
+    """
+    Return the API root page.
+
+    Provides a simple HTML page indicating the API is running.
+    Includes a noindex robots directive to prevent search engine indexing.
+
+    Returns:
+        str: HTML content describing the API.
+
+    HTTP Status Codes:
+        200: Success
+    """
     log.debug("app.py: API root page called")
     description = """
                 <!DOCTYPE html>
@@ -39,14 +82,88 @@ def hello_world():
 
 @app.route("/healthcheck", methods=["GET"])
 async def healthcheck():
-    """Route for health checking the backend"""
+    """
+    Check the health status of the OpenAlex API connection.
+
+    Verifies that the ORC can successfully communicate with the OpenAlex API.
+    This endpoint is useful for monitoring and ensuring service availability.
+
+    Returns:
+        tuple: A tuple containing:
+            - dict: Health status with keys 'healthy' (bool) and 'error' (bool or str)
+            - int: HTTP status code
+
+    HTTP Status Codes:
+        418: Healthy (OpenAlex API is accessible)
+        200: Unhealthy (OpenAlex API is not accessible or returned unexpected response)
+
+    Note:
+        The unconventional use of 418 for healthy status allows easy differentiation
+        between healthy and unhealthy states in monitoring systems.
+
+    Example Response (healthy)::
+
+        {"healthy": true, "error": false}
+
+    Example Response (unhealthy)::
+
+        {"healthy": false, "error": "Connection error"}
+    """
     log.debug("app.py: healthcheck called")
     return await orc.health_check()
 
 
 @app.route("/start_processing", methods=["POST"])
 async def start_processing():
-    """Route for returning the OpenAlex WorkIDs to the frontend"""
+    """
+    Convert a list of DOIs to OpenAlex IDs.
+
[…]

[…]
