### 2026-01-28 08faab1
https - secure connection error note
