### 2026-02-10 facec80
documentation and counter
