+++ Makefile
@@ -89,6 +89,9 @@ test_javascript:
 certificates_dry_run:
 	docker compose run --rm  certbot certonly --webroot --webroot-path /var/www/certbot/ --dry-run -d orc-demo.gesis.org
 
+certificates_renew:
+	docker compose run --rm certbot renew --webroot --webroot-path /var/www/certbot/ -n
+
 certificates_create_and_load:
 	docker compose run --rm  certbot certonly --webroot --webroot-path /var/www/certbot/ -d orc-demo.gesis.org
 
