+++ src/orc/backend/orc_backend/open_research_converter.py
@@ -96,6 +96,7 @@ class OpenResearchConverter(OpenAlexRequester):
             - submitted_count: Total DOIs submitted
             - found_count: DOIs successfully matched in OpenAlex
             - missing_dois: List of DOIs not found in OpenAlex
+            - invalid_dois: List of input strings that failed DOI format validation
         """
         new_job_id = uuid.uuid4().__str__()
         self._logger.debug(f"orc.py: new job created with id: {new_job_id}")
@@ -115,6 +116,7 @@ class OpenResearchConverter(OpenAlexRequester):
         self._jobs[new_job_id]["submitted_count"] = 0
         self._jobs[new_job_id]["found_count"] = 0
         self._jobs[new_job_id]["missing_dois"] = []
+        self._jobs[new_job_id]["invalid_dois"] = []
         return new_job_id
 
     def _recieve_data(self, job_id: str, data: list[str], email: str):
@@ -157,7 +159,8 @@ class OpenResearchConverter(OpenAlexRequester):
         Validate all input data for a processing request.
 
         Process Flow Step 11: Validates job ID exists, email is present and valid,
-        and DOIs are present and correctly formatted.
+        and DOIs are present and correctly formatted. Incorrectly formatted DOIs
+        are separated out and stored as invalid_dois; valid DOIs proceed to processing.
 
         Performs comprehensive validation of job_id, email, and DOI data
         to ensure the request can be processed.
@@ -168,15 +171,17 @@ class OpenResearchConverter(OpenAlexRequester):
             email: Email address string.
 
         Returns:
-            bool: True if all validations pass, False otherwise.
+            bool: True if job_id and email are valid and at least one valid DOI exists,
+                False otherwise.
 
         Note:
-            All three validations (job_id, email, data) must pass for the
-            overall validation to succeed.
+            Job ID and email must be valid. For DOI data, the method now partitions
+            into valid and invalid DOIs rather than rejecting the entire request.
+            Invalid DOIs are stored in the job's invalid_dois field.
         """
         job_id_is_valid = False
         email_is_valid = False
-        data_is_valid = False
+        data_has_valid = False
         if job_id is not None:
             job_id_is_valid = self._validate_uuid(job_id)
             self._logger.debug(f"job_id: {job_id}: job:{job_id_is_valid}")
@@ -184,11 +189,16 @@ class OpenResearchConverter(OpenAlexRequester):
             email_is_valid = self._validate_email(job_id, email)
             self._logger.debug(f"job_id: {job_id}: email:{email_is_valid}")
         if data is not None:
-            data_is_valid = self._validate_data(job_id, data)
-        else:
-            data_is_valid = False
-        self._logger.debug(f"job_id: {job_id}: data:{data_is_valid}")
-        return job_id_is_valid and email_is_valid and data_is_valid
+            valid_dois, invalid_dois = self._partition_dois(job_id, data)
+            self._jobs[job_id]["invalid_dois"] = invalid_dois
+            if len(invalid_dois) > 0:
+                self._logger.info(f"job_id: {job_id}: {len(invalid_dois)} invalid DOIs separated out")
+            data_has_valid = len(valid_dois) > 0
+            # Replace data in-place so _recieve_data stores only valid DOIs
+            data.clear()
+            data.extend(valid_dois)
+        self._logger.debug(f"job_id: {job_id}: data_has_valid:{data_has_valid}")
+        return job_id_is_valid and email_is_valid and data_has_valid
 
     def _validate_uuid(self, job_id: str) -> bool:
         """
@@ -303,6 +313,53 @@ class OpenResearchConverter(OpenAlexRequester):
             self._logger.error(err)
             return False
 
+    def _partition_dois(self, job_id: str, data: list[str]) -> tuple[list[str], list[str]]:
[…]
