### 2026-02-10 c9d0091
update makefile for cross platform compatibility
