### 2026-02-10 3274940
Add docker paper compile back in
