### 2026-02-11 #8205 @jhculb
Hi everyone, I think all comments have been addressed. Thank you all again for your patience while I was ill.

@evamaxfield

- There are process flow steps in the readme and corresponding in the API
- There is a counter of how many DOIs were successfully found
- I have added a section for addressing CORS issues - It is designed to be run in docker but I think the fix should work for running locally.
- I have added a makefile script that should configures the .env files automatically, with added platform independence.

@dmnapolitano

- I've added API docstrings and expanded the Readme for the case of programmatic use
- Counter added as above - including which DOIs not found
- Edits to the paper accepted and incorporated

@sarahsid10

- The makefile script should help with docker installations.
