### 2026-03-11 #8205 @dmnapolitano
Hi @jhculb ! Hope you're doing well. I'm working through my review and following your installation instructions. I managed to get the API installed and now have it running at http://127.0.0.1:8001/. When I `GET` that endpoint I get a 200 status and the message "ORC API using quart". However, if I try to `GET` http://127.0.0.1:8001/api, I get a 404 error. Did I miss something in the instructions? 👀  Thanks!
