### 2026-03-30 #8205 @logological
@jhculb Would you be able to respond to @dmnapolitano's query above?

Since we haven't heard from @evamaxfield and @sarahsid10 for a while now, I've tried pinging them by e-mail.  Hopefully the installation issue has been (or can soon be) resolved for everyone so that the remaining reviewing item checklists can be handled.
