### 2026-03-30 #8205 @evamaxfield
@logological I just saw your email right as I was testing these changes! My apologies for the great delay in resolving this!!

The documentation is greatly improved and I can now get the frontend up and running and the backend up and running. That said, I can't seem to get responses from the backend back to the frontend when I use a small test CSV file. I can see in the API server the logs are occurring as a request is sent but nothing seems to get sent back. I am wondering if this may be because OpenAlex recently changed their API to move away from polite pool and now uses API keys.

Let me know what we should do here. I have resolved everything but my last installation issue but I am happy to resolve that too as I think I am simply missing something potentially.
