### 2026-03-31 #8205 @jhculb
Hi @logological @dmnapolitano @evamaxfield, thank you for your responses.

I shall be updating the ORC ASAP (end of week) to handle the new OpenAlex API key requirements and hopefully that should fix the lack of responses.

Apologies for the delay, I have been on secondment and away from my accounts.
