### 2026-04-28 a3df11b
Adding API keys (#15)
* Fix input data processing for string with trailing comma

* Added api key update
