### 2026-04-28 #8205 @jhculb
@logological Apologies for the delay, have been on secondment. Updates now on the JOSS-1 and release branches to enable an API KEY environment variable.
