### 2026-05-01 #8205 @sarahsid10
Hi @jhculb, thank you for the updated instructions! I was able to build the Docker file and get it running. I'm testing with a small set of DOIs but it is not giving any outputs. Not sure if this has something to do with the new API KEY variable. If yes, can you further update the Docker installation steps?
I tried the same thing on the online demo and it returns the OpenAlex IDs and the bibliographic record on downloading.
