### 2026-05-20 4d1c1ec
Construct blueprint for API routes + other config changes (#16)
