### 2026-06-11 f23955d
Update readme with OpenAlex key & update key handling
