### 2026-06-15 ba50a5d
Local non-docker fix
