### 2026-06-30 2cc778f
Update Expected API response in healthcheck and remove email field
