### 2026-06-30 a1c65a9
Update BMBF to BMFTR + makefile all to test JS
