### 2026-06-30 17c6cc3
OpenAlex API version check
