### 2026-07-08 48cf6df
Remove email processing & update documentation
