+++ README.md
@@ -61,9 +61,7 @@ We provide here in the Open Research Coverter a tool utilising the OpenAlex API 
 ### Online
 If you wish to use the ORC without installing locally:
 1. Navigate to https://orc-demo.gesis.org
-2. Fill your the email address into the email box
-	* This is so that OpenAlex can monitor traffic, and places your requests in the "polite pool", where responses are faster and more consistent.
-3. Input your DOI data:
+2. Input your DOI data:
 	* The ORC expects a comma separated list of DOIs in the text box
 	* The ORC does not mind whether DOIs are prefaced with "https://doi.org/"
 	1. Via csv file
@@ -73,9 +71,9 @@ If you wish to use the ORC without installing locally:
 	2. Via copy and paste into the text box
 		* You can also manually copy and paste your DOI data into the text box
 	* The ORC can accept thousands of DOIs, though this may take a few minutes.
-4. Click Submit
+3. Click Submit
 	* A waiting animation should play in the right hand output box, if this flashes and then disappears your query may have been unsuccessful. Please try one more time, and then check your input.
-5. Wait for Output
+4. Wait for Output
 	* If your query is successful, then in the output box the first 50 OpenAlex IDs corresponding to your DOIs will be returned.
 	* If you have more submitted than 50 DOIs, then click "download CSV" to download a csv file with the DOI in the first column and the corresponding OpenAlex ID in the second column.
 ### Local Installation (Docker)
@@ -143,7 +141,7 @@ export OPENALEX_API_KEY=your-api-key  # Linux/macOS
 # Run the backend server on port 8001
 poetry run python -m quart --app src.orc.backend.orc_backend.app run --port 8001
 ```
-If successful, browsing to http://127.0.0.1:8001 should present a plaintext webpage with the following text: "ORC API using quart".
+If successful, browsing to http://127.0.0.1:8001 should present an HTML page with the heading: "ORC API using quart".
 
 **Frontend Setup:**
 ```bash
@@ -183,7 +181,7 @@ The ORC exposes a REST API for programmatic access. Full OpenAPI specification i
 
 | Method | Endpoint | Description |
 |--------|----------|-------------|
-| GET | `/api/` | API information page |
+| GET | `/` | API information page (served directly at http://host:8001/) |
 | GET | `/api/healthcheck` | Check OpenAlex API connectivity (returns 418 if healthy) |
 | POST | `/api/start_processing` | Convert DOIs to OpenAlex IDs |
 | POST | `/api/process_all` | Convert DOIs and get full OpenAlex metadata |
@@ -193,7 +191,7 @@ The ORC exposes a REST API for programmatic access. Full OpenAPI specification i
 ```bash
 curl -X POST https://orc-demo.gesis.org/api/start_processing \
   -H "Content-Type: application/json" \
-  -d '{"email": "your@email.com", "input_data": "10.1038/nature12373, 10.1126/science.1231143"}'
+  -d '{"input_data": "10.1038/nature12373, 10.1126/science.1231143"}'
 ```
 
 ### Response Format
@@ -223,16 +221,15 @@ This section describes the complete flow from when a user submits DOIs to when r
 ```
 ┌─────────────────────────────────────────────────────────────────────────────┐
 │                              USER INTERFACE                                 │
-│  1. User enters email and DOIs (via text input or CSV upload)               │
+│  1. User enters DOIs (via text input or CSV upload)                         │
 │  2. User clicks "Submit"                                                    │
 └─────────────────────────────────────────────────────────────────────────────┘
                                       │
                                       ▼
 ┌─────────────────────────────────────────────────────────────────────────────┐
 │                           FRONTEND (React)                                  │
-│  3. Validates email format (regex check)                                    │
-│  4. Sends POST request to /api/start_processing with email and DOI list     │
[…]
