### 2026-07-08 a0f4fef
Fixing healthcheck
