### 2026-07-10 a7409c4
Remove expected list response from API
