+++ paper/paper.bib
@@ -25,7 +25,7 @@
                   Bausenwein, Thomas and
                   Stahlschmidt, Stephan},
   title        = {The Data Infrastructure of the {German}
-                   {Kompetenznetzwerk} {Bibliometrie}: An Enabling
+                   \emph{{Kompetenznetzwerk} {Bibliometrie}}: An Enabling
                    Intermediary between Raw Data and Analysis},
   month        = oct,
   year         = 2024,
