+++ paper/paper.bib
@@ -10,6 +10,8 @@
 	year = {2024},
 	langid = {english},
 	keywords = {Information technology, Phase transitions and critical phenomena, Policy, Science, technology and society},
+        volume = 8,
+        number = 3,
 }
 
 @misc{schmidt_2024,
@@ -22,9 +24,9 @@
                   Taubert, Niels and
                   Bausenwein, Thomas and
                   Stahlschmidt, Stephan},
-  title        = {{The Data Infrastructure of the German
-                   Kompetenznetzwerk Bibliometrie: An Enabling
-                   Intermediary between Raw Data and Analysis}},
+  title        = {The Data Infrastructure of the {German}
+                   \emph{{Kompetenznetzwerk} {Bibliometrie}}: An Enabling
+                   Intermediary between Raw Data and Analysis},
   month        = oct,
   year         = 2024,
   publisher    = {Zenodo},
@@ -49,13 +51,13 @@
 }
 
 @Article{massimo_2024,
-  title = {openalexR: An R-Tool for Collecting Bibliometric Data from OpenAlex},
+  title = {{openalexR}: An {R}-Tool for Collecting Bibliometric Data from {OpenAlex}},
   author = {Aria Massimo and {Le Trang} and Cuccurullo Corrado and Belfiore Alessandra and Choe June},
   journal = {The R Journal},
   year = {2024},
   doi = {10.32614/RJ-2023-089},
   volume = {15},
-  issue = {4},
+  number = {4},
   issn = {2073-4859},
   pages = {167-180},
 }
@@ -64,8 +66,7 @@
   author = {Filipi Nascimento Silva},
   title = {OpenAlex Networks (openalexnet)},
   year = {2023},
-  publisher = {GitHub},
-  journal = {GitHub repository},
+  howpublished = {GitHub repository},
   url = {https://github.com/filipinascimento/openalexnet}
 }
 
@@ -73,7 +74,7 @@
   author       = {Smirnova, Nina and
                   Culbert, Jack H. and
                   Mayr, Philipp},
-  title        = {Indo-German Literature Dataset},
+  title        = {{Indo}-{German} Literature Dataset},
   month        = feb,
   year         = 2024,
   publisher    = {Zenodo},
@@ -97,7 +98,7 @@
 
 @article{gupta_2024,
 author = {Gupta, Solanki and Kanaujia, Anurag and Lathabai, Hiran H. and Singh, Vivek Kumar and Mayr, Philipp},
-title = {Patterns in the Growth and Thematic Evolution of Artificial Intelligence Research: A Study Using Bradford Distribution of Productivity and Path Analysis},
+title = {Patterns in the Growth and Thematic Evolution of Artificial Intelligence Research: A Study Using {Bradford} Distribution of Productivity and Path Analysis},
 journal = {International Journal of Intelligent Systems},
 volume = {2024},
 number = {1},
@@ -110,10 +111,13 @@ year = {2024}
 }
 
 @article{mir_2024,
-author = {Mir, A.A. and Smirnova, N. and Jeyshankar, R. and Mayr, P.},
-title = {The rise of Indo-German collaborative research: 1990–2022},
+author = {Mir, A. A. and Smirnova, N. and Jeyshankar, R. and Mayr, P.},
+title = {The rise of {Indo}-{German} collaborative research: 1990–2022},
 journal = {Global Knowledge, Memory and Communication},
 doi = {10.1108/GKMC-09-2023-0328},
 url = {https://www.emerald.com/insight/content/doi/10.1108/gkmc-09-2023-0328/full/html},
-year = {2024}
+year = {2026},
+volume = 75,
+number = {1--2},
+pages = {412--436},
 }
