### 2026-07-24 1a82e14
Citation updates
