+++ paper/paper.bib
@@ -14,23 +14,16 @@
         number = 3,
 }
 
-@misc{schmidt_2024,
-  author       = {Schmidt, Marion and
-                  Rimmert, Christine and
-                  Stephen, Dimity and
-                  Lenke, Christopher and
-                  Donner, Paul and
-                  Gärtner, Simone and
-                  Taubert, Niels and
-                  Bausenwein, Thomas and
-                  Stahlschmidt, Stephan},
-  title        = {The Data Infrastructure of the {German}
-                   \emph{{Kompetenznetzwerk} {Bibliometrie}}: An Enabling
-                   Intermediary between Raw Data and Analysis},
-  month        = oct,
-  year         = 2024,
-  publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.13932928},
+@article{schmidt2025data,
+  author       = {Schmidt, Marion and Rimmert, Christine and Stephen, Dimity and Lenke, Christopher and Donner, Paul and Gärtner, Simone and Taubert, Niels and Bausenwein, Thomas and Stahlschmidt, Stephan},
+  title        = {The Data Infrastructure of the {German} \emph{{Kompetenznetzwerk} {Bibliometrie}}: An Enabling Intermediary Between Raw Data and Analysis},
+  journal      = {Quantitative Science Studies},
+  year         = 2025,
+  volume       = 6,
+  month        = 10,
+  pages        = {1129--1146},
+  issn         = {2641-3337},
+  doi          = {10.1162/QSS.a.20},
 }
 
 
