+++ paper/paper.md
@@ -21,9 +21,9 @@ authors:
 affiliations:
   - index: 1
     name: GESIS -- Leibniz Institute for the Social Sciences
-    ror: 00hx57361
+    ror: 018afyw53
   - index: 2
-    name: Heinrich Heine Universität Düsseldorf
+    name: Heinrich Heine University Düsseldorf
     ror: 024z2rq82
 date: 5 November 2024
 bibliography: paper.bib
