+++ CITATION.cff
@@ -34,3 +34,5 @@ keywords:
   - DOI
   - Bibliometrics
 license: GPL-3.0-or-later
+version: 1.2.0
+date-released: '2026-07-27'
