+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "Open Research Converter"
-version = "1.1.1"
+version = "1.2.0"
 authors = [
     "Jack Culbert <jack.culbert@gesis.org>",
     "Ahsan Shahid <ahsan.shahid@gesis.org>",
