### 2026-07-27 ab6e93e
Updated text - size of OAL, API key requirement.
