+++ CITATION.cff
@@ -1,7 +1,7 @@
 # This CITATION.cff file was generated with cffinit.
 # Visit https://bit.ly/cffinit to generate yours today!
 
-cff-version: 1.2.0
+cff-version: 1.2.1
 title: Open Research Converter
 message: >-
   If you use this software, please cite it using the
@@ -34,5 +34,5 @@ keywords:
   - DOI
   - Bibliometrics
 license: GPL-3.0-or-later
-version: 1.2.0
+version: 1.2.1
 date-released: '2026-07-27'
